Title: 숫자 사이의 배수, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240127/숫자 사이의 배수/multiplication-between-numbers.c
@@ -15,7 +15,7 @@ int main() {
             sum += i;
         }
     }
-
-    printf("%d %.1f",sum, sum/cnt);
+    
+    printf("%d %.1f",sum, (float)sum/cnt);
     return 0;
 }
